Remove PlaceManager
