+++ uberfire-workbench/uberfire-workbench-client/src/main/java/org/uberfire/client/mvp/ActivityManager.java
@@ -19,6 +19,8 @@ import org.uberfire.mvp.PlaceRequest;
 
 public interface ActivityManager {
 
+    Activity getEditorActivity();
+
     Activity getActivity(final PlaceRequest placeRequest);
 
     void openActivity(final String activityId);
